+++ application/views/common-js/mega-menu.js
@@ -1,8 +1,15 @@
 $(function () {
+    var isOnDiv = false;
+    $('.navchild').mouseenter(function () { isOnDiv = true; });
+    $('.navchild').mouseleave(function () { isOnDiv = false; });
 
     /* for navigation drop down */
     $('.navchild').hover(function () {
         var el = $("body");
+        if (isOnDiv === true) {
+            $(this).removeClass('is-focus');
+            el.removeClass('nav_show');
+        }
         if ($(window).width() > 1025) {
             if ($(this).hasClass('is-focus')) {
                 $(this).removeClass('is-focus');