+++ CHANGELOG.md
@@ -1144,6 +1144,8 @@ TV-9.4.0.20220719
     - #065950 - The download count does not get increased if the admin / seller attach any file with existing digital product order
     - #066106 - The product is visible on the front end even if the tax category is not bound with the product.
     - #066101 - Nothing happens on clicking on the 'Write A Review' button on the reviews permalink page
+    - #066017 - Record not appearing in subscription report by seller
+    - #066015 - Mutiple issues in report for SUbscription
     - #066073 - The search field on the categories listing page is not working properly
     - #066132 - Parse error occurs when the admin tries to delete the category thumb.